refactor(types): add DockiteFieldStatic type to base types package
Add the DockiteFieldStatic type to the base types package to resolve some
potential circular dependencies.
