+++ packages/types/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "express": "^4.17.1",
     "graphql": "^14.6.0",
+    "typeorm": "^0.2.22",
     "typescript": "~3.7.5"
   },
   "devDependencies": {
