+++ packages/types/src/common/field.ts
@@ -3,15 +3,30 @@ import {
   GraphQLOutputType,
   GraphQLObjectType,
   GraphQLFieldConfigArgumentMap,
+  GraphQLSchema,
 } from 'graphql';
+import { Repository } from 'typeorm';
 
-import { Schema } from '../entities';
+import { Schema, Field } from '../entities';
 
 export interface FieldContext {
   value: any;
   root: Record<string, any>;
   args: Record<string, any>;
   context: Record<string, any>;
+}
+
+export interface DockiteFieldStatic {
+  type: string;
+  title: string;
+  description: string;
+  defaultOptions: object;
+
+  new (
+    schemaField: Field,
+    repositories: { [id: string]: Repository<any> },
+    schema: GraphQLSchema | null,
+  ): DockiteField;
 }
 
 export interface DockiteField {
@@ -24,6 +39,7 @@ export interface DockiteField {
   outputType(
     dockiteSchemas: Schema[],
     types: Map<string, GraphQLObjectType>,
+    dockiteFields: Record<string, DockiteFieldStatic>,
   ): Promise<GraphQLOutputType>;
 
   outputArgs(): Promise<GraphQLFieldConfigArgumentMap>;
